tmf: Distinguish between simple regex and complex table column search
Simple regex search allows clients to provide a single search string
per column. This will use regex matches on the column content.

The complex filter string allows the for free text form according
to the org.eclipse.tracecompass.tmf.filter.parser plug-in.

Provide the filter expression in the fetch parameter in a map
(column ID -> filter expression) which will allow to mark the cells
as matched cells in a later version of the data provider.

[Added] Simple regex and complex table column searches for data provider UIs

Change-Id: I9b80aa04fa40c2190da514d7334fc293ae816ebc
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/181356
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>