tmf: handle fetch parameter requested_marker_categories in data provider
All Trace Compass defined data providers handle the fetch parameter
requested_marker_categories in the implementation of the data provider
method fetchAnnotations() to filter at the source.

If requested_marker_categories is not part of the fetch parameters,
then all the markers are returned. If it present, the all markers
in the requested list of categories are returned. Use empty list, to
filter out all of them.

[Changed] handle requested_marker_categories in data provider
[Added] DataProviderParameterUtils#extractSelectedCategories

Change-Id: I0ea43dd9b217225160c223c9a34bf07e6a6b92c3
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/182061
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>