lttng: Add benchmark for critical path computation
The KernelExecutionGraphBenchmark is updated to also benchmark the
critical path computation of a specific thread. Assertions are also made
on the number of workers in the graph and critical path.

The benchmark is parameterized and takes the path of the trace, so one
can easily add their own traces to better benchmark various local
traces.

Change-Id: I37decc664acf10350dad5f78ec4fd31cce470a0e
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/182710
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>