tmf: Make only one CustomOutputAnnotationProvider return annotations
The CustomOutputAnnotationProvider instances are associated with a
trace/experiment by a constructor parameter.

When annotations are being fetched, only the provider associated with
the 'requested_trace' fetch parameter will respond with a model. The
other providers will return a null model. If the 'requested_trace'
parameter is absent, only the provider associated with the first element
of the opened trace/experiment trace set will return a model. The other
providers will return a null model.

When annotation categories are being fetched, only the provider
associated with the first element of the opened trace/experiment trace
set will return a model. The other providers will return a null model.

The 'requested_trace' fetch parameter is no longer included when
fetching annotation categories, it is not required. When fetching
annotations, it is not included to allow for the default (first trace)
to be used, as there is currently no UI control to select a requested
trace.

The 'requested_marker_set' and 'requested_trace' fetch parameters are
removed from the requests sent to specific view output annotation data
providers. These parameters are only relevant to the
CustomOutputAnnotationProvider trace adapter output annotation provider.

The classes CustomDefinedOutputAnnotationProvider/Factory are renamed to
CustomOutputAnnotationProvider/Factory

[Fixed] Only one CustomOutputAnnotationProvider return annotations
[Fixed] Remove unnecessary parameters when fetching annotations
[Changed] CustomDefinedOutputAnnotationProvider/Factory renamed

Change-Id: I89ecf729fa958532ad6d80bbe85cae8ee5b8824d
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/182716
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>