tmf: Line search should start at requested table index
When a search filter is provided in a fetchLine() requested, the search
forward or backward should start at the requested table index, not at
the next or previous index.

If the index of the current selection should be skipped, it should be
the responsibility of the client to adjust the requested index.

[Fixed] Line search should start at requested table index

Change-Id: I0e8e2152fda2b1c259b9affdb7a2d593b5538158
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/183874
Reviewed-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>