ctf: merge packet context into event content
The (stream-specific) packet context may contain additional
context information which is constant for every event. This is typically
a good place to store large strings. In order to make these available in each
event's content as 'context.*' fields, the packet context must be merged into
the content as it is already done with the stream event context and event
context.

Bug 575394

Change-Id: I97b5380972a8ff875245e795e3b3a058865e580c
Signed-off-by: Johannes Schlatow <johannes.schlatow@genode-labs.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/188748
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>