Bug 576647: analysis: Add WaitForCompletion for segment store analysis
This makes sure that the segment store analysis is finished and the
fetchTree() doesn't return status FAILED when the analysis is running.

The fetchTree should return RUNNING instead but this require an API
addition to support that. This patch is intermediate fix until the
new API is available and can be used in this data provider.

Change-Id: I9052ea606b9e993d275fc0af6ff3e85d5e9a4c9e
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/186536
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>