segmentstore: rename references to SegmentStoreProvider
Before this commit, AbstractSegmentStatisticsAnalysis
methods and fields related to SegmentStoreProvider
were using the terms "module" or "analysis". This was
suggesting that a segment store provider is supposed
to be an analysis. A segment store provider can be
an analysis, but it does not have to.

After this commit, AbstractSegmentStatisticsAnalysis
fields and method are renamed to remove references
to "modules" or "analysis", hopefully clarifying that
a segment store provider does not need to be
an analysis module.

[Added] AbstractSegmentStatisticsAnalysis#getSegmentStoreProvider
[Deprecated] AbstractSegmentStatisticsAnalysis#getSegmentProviderAnalysis

Signed-off-by: Francesco Robino <francesco.robino@ericsson.com>
Change-Id: Ia92d3e2cd3fc6e87d5e3fd94ace1e5085c329bc1
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/193127
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>