Change tooltip background for sequence diagram
Before this change, the text in the tooltip displayed
in the sequence diagram was not visible since the background
color is the same as the text color. This change makes the
tooltip background transparent so the text can be seen by
removing the unnecessary background setter for the text box and by
removing the unnecessary background setter for the Tooltip shell
from drawable class.

Change-Id: I13e810cb1c998aa37152214c1978dbb9ed9c101c
Signed-off-by: Kyrollos Bekhet <kyrollos.bekhet@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/193596
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>