timing.core: Segment Store data provider for tables
Since the segment store can't be sent over the network to be shown in a
table, a data provider has to be implemented to fetch the desired number
of lines starting from an index from a given segment store to be able to
display the data.

Since the segment store is not index based, an indexing mechanism is
implemented to match the index given in the request with the segment in
the segment store. The indexes are stored in an array that can use an
index to extract the segment index which has a start and end timestamp.

Adding a factory that can be used by theia-trace-extension to create a
segment store table to show the data in the segments.

[Added] Dataprovider for Segment store table

Change-Id: Id83b8a4375602de569c266a8e8ad6fe1a0495fcd
Signed-off-by: Kyrollos Bekhet <kyrollos.bekhet@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/193950
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>