statesystem: Extract HistoryTreeBackend{Iterator}
Extract an HistoryTreeBackendIterator class from the previous, anonymous
class implementation. Enable the use of IDE breakpoints within e.g.
(now HistoryTreeBackendIterator's) hasNext, this way, for manual or test
debugging purposes.

Pave the way for potentially adding finer log trace events that way too,
for e.g. tracing performance troubleshooting. However preserve the
previous implementation to the letter for this change scope, beside the
necessary yet trivial anonymity undoings.

Reuse the HistoryTreeBackend file header and history lines as is for
this HistoryTreeBackendIterator, but add a line about this specific
change of having extracted the latter. Also add 2022 as this very year.

[Added] HistoryTreeBackendIterator as extracted from HistoryTreeBackend

Change-Id: I741c6c46db4b2e80ae1fcc4545a17097bfd23dc0
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/194434
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>