statesystem: Extract {HTNode}CacheLoader as nested
Extract an HTNodeCacheLoader out of the existing yet overridden
CacheLoader instance, as an HT_IO static nested class, consistently with
the previous implementation. Use an instance of this class as before.

Make this clean while preparing the ground for an upcoming addition,
which will require such a class of its own.

Change-Id: I9a0e7d09a07c0f0633660b1c398b8978e7c5ce1c
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/194552
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>