tmf.ui: Make factor effective in BDPTimeGraphView
Sonarlint warned about this local variable being reassigned with no
effect. Make such potential reassignments effective with this move,
which was likely the original intention anyway when writing those lines.

This change only makes the sleep longer across each potential iteration,
while capping that wait factor if everlasting. Should this behaviour no
longer be the wanted one, then this factor should be replaced with a
hard-coded or constant 1.0 value; this was the behaviour before.

Change-Id: I5687cbb14ea36ca025639d308ead274cc6800e79
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/194577
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>